rework locationmanager & add OOP

+++ src/main/java/com/panda0day/bedwars/location/LocationManager.java
@@ -1,4 +1,4 @@
-package com.panda0day.bedwars.utils;
+package com.panda0day.bedwars.location;
 
 import com.panda0day.bedwars.Main;
 import org.bukkit.Bukkit;
@@ -6,49 +6,57 @@ import org.bukkit.Location;
 import org.bukkit.World;
 
 import java.sql.ResultSet;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 public class LocationManager {
-    public static Location getLocation(String name) {
+    private final List<Locations> locationsList;
+
+    public LocationManager() {
+        this.locationsList = new ArrayList<>();
+    }
+
+    public void loadLocations() {
         try {
-            ResultSet resultSet = Main.getDatabase().executeQuery("SELECT * FROM locations WHERE name = ?;", name);
-            while (resultSet.next()) {
-                String worldName = resultSet.getString("world");
+            ResultSet rs = Main.getDatabase().executeQuery("SELECT * FROM locations");
+            while (rs.next()) {
+                String name = rs.getString("name");
+                String worldName = rs.getString("world");
                 if (!Main.getWorldManager().isWorldLoaded(worldName)) {
                     Main.getWorldManager().loadWorld(worldName);
                 }
                 World world = Bukkit.getWorld(worldName);
-                double x = resultSet.getDouble("x");
-                double y = resultSet.getDouble("y");
-                double z = resultSet.getDouble("z");
-                float yaw = resultSet.getFloat("yaw");
-                float pitch = resultSet.getFloat("pitch");
-                return new Location(world, x, y, z, yaw, pitch);
+                double x = rs.getDouble("x");
+                double y = rs.getDouble("y");
+                double z = rs.getDouble("z");
+                float yaw = rs.getFloat("yaw");
+                float pitch = rs.getFloat("pitch");
+                this.locationsList.add(new Locations(name, new Location(world, x, y, z, yaw, pitch)));
             }
         } catch (Exception exception) {
             Main.getInstance().getLogger().info(exception.getMessage());
         }
-
-        return null;
     }
 
-    public static boolean doesLocationExist(String name) {
-        try {
-            ResultSet resultSet = Main.getDatabase().executeQuery("SELECT * FROM locations WHERE name = ?;", name);
-            return resultSet.next();
-        } catch (Exception exception) {
-            Main.getInstance().getLogger().info(exception.getMessage());
-        }
-
-        return false;
+    public Locations getLocation(String name) {
+        return this.locationsList.stream().findFirst().filter(x -> x.getName().equalsIgnoreCase(name)).orElse(null);
     }
 
-    public static boolean deleteLocation(String name) {
+    public boolean doesLocationExist(String name) {
+        return this.locationsList.stream().findFirst().filter(x -> x.getName().equalsIgnoreCase(name)).orElse(null) != null;
+    }
+
+    public boolean deleteLocation(String name) {
+        Locations location = this.getLocation(name);
+        if (location == null) return false;
+        this.locationsList.remove(location);
+
         ResultSet resultSet = Main.getDatabase().executeQuery("DELETE FROM locations WHERE name = ?;", name);
         return resultSet == null;
     }
 
-    public static boolean setLocation(String name, Location location) {
+    public boolean setLocation(String name, Location location) {
         ResultSet resultSet = Main.getDatabase().executeQuery("INSERT INTO locations (name, world, x, y, z, yaw, pitch) VALUES (?, ?, ?, ?, ?, ?, ?);",
                 name,
                 Objects.requireNonNull(location.getWorld()).getName(),
@@ -59,10 +67,12 @@ public class LocationManager {
                 location.getPitch()
         );
 
+        this.locationsList.add(new Locations(name, location));
+
         return resultSet == null;
     }
 
-    public static boolean areLocationsEqual(Location location1, Location location2) {
+    public boolean areLocationsEqual(Location location1, Location location2) {
         if (location1 == null || location2 == null) {
             return false;
         }
